refactor: Embed traQ client tests within `traq_client.rs`
