+++ crates/infra/src/traq_client.rs
@@ -185,3 +185,395 @@ impl TraqClient for TraqClientImpl {
         Ok(message)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use oauth2::{
+        AuthUrl, AuthorizationCode, ClientId, ClientSecret, CsrfToken, Scope, TokenResponse,
+        TokenUrl, basic::BasicClient,
+    };
+    use std::time::Duration;
+    use testcontainers::compose::DockerCompose;
+    use uuid::Uuid;
+
+    /// Test environment that orchestrates traQ via Docker Compose
+    struct TraqTestEnvironment {
+        compose: Option<DockerCompose>,
+        base_url: String,
+        admin_token: Option<String>,
+        admin_user_id: Option<Uuid>,
+    }
+
+    impl TraqTestEnvironment {
+        async fn start() -> anyhow::Result<Self> {
+            // Get workspace root (crates/infra/src -> project root)
+            let workspace_root = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+                .parent()
+                .unwrap()
+                .parent()
+                .unwrap()
+                .to_path_buf();
+
+            let compose_file = workspace_root.join("compose.yaml");
+
+            // Generate unique project name
+            let project_name = format!("traq_test_{}", Uuid::now_v7().simple());
+
+            // Create and configure compose instance
+            let mut compose = DockerCompose::with_local_client(&[compose_file.to_str().unwrap()])
+                .with_env("COMPOSE_PROJECT_NAME", &project_name)
+                .with_env("COMPOSE_PROFILES", "dev")
+                .with_env("TRAQ_SERVER_PORT", "0") // Random port assignment
+                .with_env("MARIADB_PORT", "0")
+                .with_env("ADMINER_PORT", "0");
+
+            // Start services
+            compose.up().await?;
+
+            // Wait for services to initialize
+            tokio::time::sleep(Duration::from_secs(10)).await;
+
+            // Get traq_server service and mapped port
+            let traq_server = compose
+                .service("traq_server")
+                .expect("traq_server service not found");
+            let port = traq_server.get_host_port_ipv4(3000).await?;
+
+            // base_url for traQ API (with /api/v3 prefix)
+            let api_base_url = format!("http://localhost:{}/api/v3", port);
+            // server_base_url for OAuth endpoints (no /api/v3)
+            let server_base_url = format!("http://localhost:{}", port);
+
+            // Initialize traQ and get admin token via OAuth2
+            let (admin_token, admin_user_id) =
+                match Self::initialize_traq_oauth(&api_base_url, &server_base_url).await {
+                    Ok(result) => (Some(result.0), Some(result.1)),
+                    Err(e) => {
+                        eprintln!("WARNING: traQ initialization failed: {}", e);
+                        (None, None)
+                    }
+                };
+
+            Ok(Self {
+                compose: Some(compose),
+                base_url: api_base_url,
+                admin_token,
+                admin_user_id,
+            })
+        }
+
+        async fn initialize_traq_oauth(
+            api_base_url: &str,
+            _server_base_url: &str,
+        ) -> anyhow::Result<(String, Uuid)> {
+            let client = reqwest::Client::builder()
+                .timeout(Duration::from_secs(10))
+                .cookie_store(true)
+                .redirect(reqwest::redirect::Policy::none())
+                .build()?;
+
+            // Wait for traQ to be ready
+            eprintln!("Waiting for traQ at {}...", api_base_url);
+            for i in 0..60 {
+                match client.get(format!("{}/version", api_base_url)).send().await {
+                    Ok(res) if res.status().is_success() => {
+                        eprintln!("traQ ready!");
+                        break;
+                    }
+                    Ok(res) => eprintln!("Status: {}", res.status()),
+                    Err(e) => eprintln!("Attempt {}/60: {}", i + 1, e),
+                }
+                if i == 59 {
+                    anyhow::bail!("traQ not ready after 60 attempts");
+                }
+                tokio::time::sleep(Duration::from_secs(1)).await;
+            }
+
+            // Login with default user (traq/traq)
+            eprintln!("Logging in with default user (traq/traq)...");
+            let login_res = client
+                .post(format!("{}/login", api_base_url))
+                .json(&serde_json::json!({
+                    "name": "traq",
+                    "password": "traq"
+                }))
+                .send()
+                .await?;
+
+            if !login_res.status().is_success() {
+                anyhow::bail!("Login failed with default user");
+            }
+
+            // Get user ID
+            let me_res = client
+                .get(format!("{}/users/me", api_base_url))
+                .send()
+                .await?;
+            let me_data: serde_json::Value = me_res.json().await?;
+            let user_id = Uuid::parse_str(me_data["id"].as_str().unwrap())?;
+
+            // Create OAuth client
+            eprintln!("Creating OAuth client...");
+            let client_res = client
+                .post(format!("{}/clients", api_base_url))
+                .json(&serde_json::json!({
+                    "name": "test_client",
+                    "description": "Test client for integration testing",
+                    "callbackUrl": "http://localhost:3000/callback",
+                    "scopes": ["read", "write"]
+                }))
+                .send()
+                .await?;
+
+            if !client_res.status().is_success() {
+                let error = client_res.text().await?;
+                anyhow::bail!("Failed to create OAuth client: {}", error);
+            }
+
+            let client_data: serde_json::Value = client_res.json().await?;
+            let client_id = client_data["id"].as_str().unwrap();
+            let client_secret = client_data["secret"].as_str().unwrap();
+
+            // Set up OAuth2 client
+            let oauth_client = BasicClient::new(ClientId::new(client_id.to_string()))
+                .set_client_secret(ClientSecret::new(client_secret.to_string()))
+                .set_auth_uri(AuthUrl::new(format!("{}/oauth2/authorize", api_base_url))?)
+                .set_token_uri(TokenUrl::new(format!("{}/oauth2/token", api_base_url))?);
+
+            // Generate authorization URL
+            let (auth_url, _csrf_state) = oauth_client
+                .authorize_url(CsrfToken::new_random)
+                .add_scope(Scope::new("read".to_string()))
+                .add_scope(Scope::new("write".to_string()))
+                .url();
+
+            // Get authorization (redirects to consent)
+            let auth_res = client.get(auth_url.as_str()).send().await?;
+
+            // Extract authorization code
+            let code = if auth_res.status().is_redirection() {
+                let location = auth_res
+                    .headers()
+                    .get("location")
+                    .and_then(|v| v.to_str().ok())
+                    .ok_or_else(|| anyhow::anyhow!("No location header"))?;
+
+                if location.contains("/consent") {
+                    // Approve via /oauth2/authorize/decide
+                    let decide_url = format!("{}/oauth2/authorize/decide", api_base_url);
+                    let approve_res = client
+                        .post(&decide_url)
+                        .form(&[("submit", "approve")])
+                        .send()
+                        .await?;
+
+                    if !approve_res.status().is_redirection() {
+                        anyhow::bail!("Consent approval failed");
+                    }
+
+                    let location = approve_res
+                        .headers()
+                        .get("location")
+                        .and_then(|v| v.to_str().ok())
+                        .ok_or_else(|| anyhow::anyhow!("No location after approval"))?;
+
+                    let callback_url = url::Url::parse(location)?;
+                    callback_url
+                        .query_pairs()
+                        .find(|(key, _)| key == "code")
+                        .map(|(_, value)| value.to_string())
+                        .ok_or_else(|| anyhow::anyhow!("No code in callback"))?
+                } else {
+                    let callback_url = url::Url::parse(location)?;
+                    callback_url
+                        .query_pairs()
+                        .find(|(key, _)| key == "code")
+                        .map(|(_, value)| value.to_string())
+                        .ok_or_else(|| anyhow::anyhow!("No code in callback"))?
+                }
+            } else {
+                anyhow::bail!("Unexpected authorization response");
+            };
+
+            // Exchange code for token
+            let http_client = reqwest::Client::new();
+            let token_result = oauth_client
+                .exchange_code(AuthorizationCode::new(code))
+                .request_async(&http_client)
+                .await?;
+
+            let access_token = token_result.access_token().secret().to_string();
+            eprintln!("Got access token!");
+
+            Ok((access_token, user_id))
+        }
+
+        fn base_url(&self) -> &str {
+            &self.base_url
+        }
+
+        fn admin_token(&self) -> &str {
+            self.admin_token.as_deref().unwrap_or("")
+        }
+
+        fn admin_user_id(&self) -> Option<Uuid> {
+            self.admin_user_id
+        }
+
+        /// Explicit cleanup method
+        async fn cleanup(mut self) -> anyhow::Result<()> {
+            if let Some(compose) = self.compose.take() {
+                eprintln!("Cleaning up Docker Compose resources...");
+                compose.down().await?;
+                eprintln!("Cleanup complete!");
+            }
+            Ok(())
+        }
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_environment_starts() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ environment");
+
+        assert!(!env.base_url().is_empty());
+        assert!(!env.admin_token().is_empty());
+        assert!(env.admin_user_id().is_some());
+
+        println!("traQ URL: {}", env.base_url());
+        println!("Admin token: {}", env.admin_token());
+        println!("Admin user ID: {:?}", env.admin_user_id());
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_get_user_success() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ");
+
+        let client = TraqClientImpl::new(env.base_url().to_string());
+        let user_id = env.admin_user_id().expect("No admin user ID");
+
+        let result = client.get_user(env.admin_token(), &user_id).await;
+
+        assert!(result.is_ok());
+        let user = result.unwrap();
+        assert_eq!(user.id, user_id);
+        assert_eq!(user.handle, "traq");
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_get_user_not_found() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ");
+
+        let client = TraqClientImpl::new(env.base_url().to_string());
+        let non_existent_id = Uuid::new_v4();
+
+        let result = client.get_user(env.admin_token(), &non_existent_id).await;
+
+        assert!(result.is_err());
+        let error_msg = result.unwrap_err().to_string();
+        assert!(error_msg.contains("404") || error_msg.contains("Not Found"));
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_get_user_unauthorized() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ");
+
+        let client = TraqClientImpl::new(env.base_url().to_string());
+        let user_id = env.admin_user_id().expect("No admin user ID");
+
+        let result = client.get_user("invalid_token", &user_id).await;
+
+        assert!(result.is_err());
+        let error_msg = result.unwrap_err().to_string();
+        assert!(error_msg.contains("401") || error_msg.contains("Unauthorized"));
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_get_stamps_success() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ");
+
+        let client = TraqClientImpl::new(env.base_url().to_string());
+
+        let result = client.get_stamps(env.admin_token()).await;
+
+        assert!(result.is_ok());
+        let stamps = result.unwrap();
+        // traQ has default stamps
+        assert!(!stamps.is_empty());
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_get_stamp_success() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ");
+
+        let client = TraqClientImpl::new(env.base_url().to_string());
+
+        // First get all stamps to get a valid ID
+        let stamps = client
+            .get_stamps(env.admin_token())
+            .await
+            .expect("Failed to get stamps");
+        assert!(!stamps.is_empty());
+
+        let stamp_id = stamps[0].id;
+
+        // Now get individual stamp
+        let result = client.get_stamp(env.admin_token(), &stamp_id).await;
+
+        assert!(result.is_ok());
+        let stamp = result.unwrap();
+        assert_eq!(stamp.id, stamp_id);
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+
+    #[tokio::test]
+    #[ignore]
+    async fn test_fetch_messages_since() {
+        let env = TraqTestEnvironment::start()
+            .await
+            .expect("Failed to start traQ");
+
+        let client = TraqClientImpl::new(env.base_url().to_string());
+
+        // Search messages from a week ago
+        let since = OffsetDateTime::now_utc() - time::Duration::days(7);
+
+        let result = client.fetch_messages_since(env.admin_token(), since).await;
+
+        assert!(result.is_ok());
+        let messages = result.unwrap();
+        // May be empty in freshly created traQ instance
+        println!("Found {} messages", messages.len());
+
+        env.cleanup().await.expect("Failed to cleanup");
+    }
+}
